requirements.txt
selenium>=4.0.0
requests>=2.28.0
allure-pytest>=2.9.0
pytest>=7.0.0
webdriver-manager>=3.5.0
python-dotenv>=0.19.0